chore: minor tweaks to websocket-proxy (#353)
* chore: remove integration feature for websocket-proxy

* chore: received message metric per upstream source

* fix: ensure we reconnect when we receive a close

* Reformat

* Add a metric to track # of bytes broadcasted
