+++ crates/infra/websocket-proxy/Cargo.toml
@@ -29,13 +29,13 @@ reqwest = { version = "0.12.15", default-features = false, features = [
 metrics-derive.workspace = true
 hostname = "0.4.0"
 redis = "0.30.0"
-redis-test = { version = "0.10.0", optional = true }
 uuid = { version = "1.16.0", features = ["v4"] }
 tokio-util = "0.7.12"
 brotli = "8.0.1"
 
-[dependencies.ring]
-version = "0.17.12"
+[lib]
+name = "websocket_proxy"
 
-[features]
-integration = ["redis-test"]
+[dev-dependencies]
+testcontainers.workspace = true
+testcontainers-modules.workspace = true
