+++ crates/infra/websocket-proxy/src/lib.rs
@@ -1,0 +1,15 @@
+pub mod auth;
+pub mod client;
+pub mod metrics;
+pub mod rate_limit;
+pub mod registry;
+pub mod server;
+pub mod subscriber;
+
+pub use auth::*;
+pub use client::*;
+pub use metrics::*;
+pub use rate_limit::*;
+pub use registry::*;
+pub use server::*;
+pub use subscriber::*;
