+++ crates/infra/websocket-proxy/src/metrics.rs
@@ -30,9 +30,6 @@ pub struct Metrics {
     #[metric(describe = "Count of times upstream receiver was closed/errored")]
     pub upstream_errors: Counter,
 
-    #[metric(describe = "Count of messages received from the upstream source")]
-    pub upstream_messages: Gauge,
-
     #[metric(describe = "Number of active upstream connections")]
     pub upstream_connections: Gauge,
 
@@ -44,10 +41,18 @@ pub struct Metrics {
 
     #[metric(describe = "Number of failed upstream connection attempts")]
     pub upstream_connection_failures: Counter,
+
+    #[metric(describe = "Total bytes broadcasted to clients")]
+    pub bytes_broadcasted: Counter,
 }
 
 impl Metrics {
     pub fn proxy_connections_by_app(&self, app: &str) {
         counter!("websocket_proxy.connections_by_app", "app" => app.to_owned()).increment(1);
+    }
+
+    pub fn message_received_from_upstream(&self, upstream: &str) {
+        counter!("websocket_proxy.upstream_messages", "upstream" => upstream.to_owned())
+            .increment(1);
     }
 }
